fix(linux): GameSpy Thread + CompatLib architecture + PCH integration (fighter19 patterns)
Applied fighter19 Linux compatibility patterns across 8 GameSpy files +
comprehensive CompatLib architecture overhaul + PCH integration.

ARCHITECTURAL CHANGES (CompatLib):

windows_compat.h:
- Added thread_compat.h include (fighter19 line 69 central hub pattern)
- Removed duplicate GetCurrentThreadId/Sleep (conflicted with thread_compat.h)
- Added FAR macro stub (legacy 16-bit Windows pointer qualifier)
Reason: Single source of truth for threading APIs, prevents ambiguous overloads

socket_compat.h:
- Removed CreateThread stub (conflicted with thread_compat.h real implementation)
- Removed LPTHREAD_START_ROUTINE typedef (use start_routine from thread_compat.h)
Reason: FTP.cpp CreateThread call became ambiguous (socket stub vs thread real)

PCH INTEGRATION (NEW):

GameEngineDevice/Include/PchCompat.h + WWLib/PchCompat.h:
- Added windows_compat.h inclusion in Linux branch
Reason: Provides __max/__min macros needed by W3DBufferManager.cpp (lines 311, 431)
Impact: All files using PCH now have full CompatLib layer

GAMESPY FILES (8 total):

GameResultsThread.cpp: winsock guards + HOSTENT→hostent
BuddyThread.cpp: (Int)param → (uintptr_t)param (64-bit safe)
PeerThread.cpp: winsock guards + int→socklen_t
LobbyUtils.cpp: prevPos declaration moved outside if block (fighter19 line 700)
MainMenuUtils.cpp: POSIX file I/O + 4× uintptr_t casts
PingThread.cpp: winsock guards + HOSTENT fix + gethostbyname call + DWORD types
IPEnumeration.cpp: HOSTENT→hostent
FTP.cpp: Compiled successfully after socket_compat.h cleanup [14/699]

BUILD PROGRESSION:
- Attempt 1 [4/124]: Fixed winsock/HOSTENT/scope/POSIX
- Attempt 2 [208/904]: thread_compat integration + FAR macro
- Attempt 3 [6/699]: CreateThread ambiguity resolved
- Attempt 4 [26/699]: PCH integration completed
- Attempt 5 [7/129]: PingThread SUCCESS!

VALIDATION:
GameSpy compatibility PRESERVED (opensource server GeneralsOnlineDevelopmentTeam).
Windows builds fully preserved behind platform guards.
Scope expanded 5→8 files during iterative debugging.

NEW BLOCKERS:
- LANAPI.cpp: WideCharWindows* vs wchar_t* (10+ errors)
- LANAPIhandlers.cpp: WideCharWindows/UnicodeString (26+ errors)
- LANAPICallbacks.cpp: void* to UnsignedInt cast (1 error)

Progress: Resolved 30+ GameSpy Thread errors + 3 cascading architecture issues

TheSuperHackers @build BenderAI 13/02/2026

+++ Core/GameEngine/Source/GameNetwork/GameSpy/Thread/BuddyThread.cpp
@@ -125,7 +125,7 @@ enum CallbackType
 
 void callbackWrapper( GPConnection *con, void *arg, void *param )
 {
-	CallbackType info = (CallbackType)(Int)param;
+	CallbackType info = (CallbackType)(uintptr_t)param;
 	BuddyThreadClass *thread = MESSAGE_QUEUE->getThread() ? MESSAGE_QUEUE->getThread() : nullptr /*(TheGameSpyBuddyMessageQueue)?TheGameSpyBuddyMessageQueue->getThread():nullptr*/;
 	if (!thread)
 		return;
